tests/wpt/web-platform-tests/url/urlsearchparams-foreach.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<link rel="help" href="http://url.spec.whatwg.org/#dom-urlsearchparams-has">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    var params = new URLSearchParams('a=1&b=2&c=3');
    for(var URLSearchParam in params){
        switch (URLSearchParam) {
            case 'a':
                assert_equals('1', params[URLSearchParam]);
                break;
            case 'b':
                assert_equals('2', params[URLSearchParam]);
                break;
            case 'c':
                assert_equals('3', params[URLSearchParam]);
                break;
        }
    }
}, "For-in Check");

test(function() {
    var params = new URLSearchParams('a=1&b=2&c=3');
    params.forEach(function(value,index){
        switch (index) {
            case 'a':
                assert_equals('1', value);
                break;
            case 'b':
                assert_equals('2', value);
                break;
            case 'c':
                assert_equals('3', value);
                break;
        }
    });
}, "ForEach Check");
</script>
</head>
</html>
